+++ src/infrastructure/models/payment_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, DateTime
+from sqlalchemy import Column, ForeignKey, Integer, String, DateTime, Numeric
 from infrastructure.databases.base import Base
 from datetime import datetime
 class Payment(Base):
@@ -8,7 +8,7 @@ class Payment(Base):
     id = Column(Integer, primary_key=True) # Cho phép NULL
     invoice_id=Column(Integer, ForeignKey("invoices.id"),nullable=False)
     method_id=Column(Integer,ForeignKey("paymentmethods.id"),nullable=False)
-    amount=Column(Integer, nullable=False)
+    amount=Column(Numeric(10), nullable=False)
     description = Column(String(255), nullable=True)
     status = Column(String(50), nullable=False)
     created_by= Column(String(50),nullable=True)
